refactor: Move binary files definition to ggshield.utils
The list of extensions that are considered as binary files is not
specific to ggshield and should be in the `utils` packages according to
the new architecture.

refs #521
