+++ ggshield/utils/_binary_extensions.py
